Move `AggregationFormula` into its own module
